Split off webworker / node compatibility into separate file
